Allow retry of path during hmac validation for a8
Temporary: A8 places the service name in the path, which leads to the path check failing the hmac because paths are sent to a8 as /map/map/fish but then the call is placed onwards to the service as /map/fish. So the hmac was calculated with /map/map/fish originally by the client, and thus the check fails when the service validates with only /map/fish. This change adds the first path segment back to the path, then revalidates the hmac if it failed initially. 
Once we rewrite the nginx config for a8, this will no longer be required..